Updated folder structure of Patients
